backend/requirements.txt
python-dotenv
fastapi
uvicorn
python-multipart
ultralytics
torch
torchvision
pillow
numpy
google-generativeai